move dbt-rpc related code to separate class
